+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Changelog
 
 * Fixed all minor bugs reported through the support info
 * Fixed some minor bugs and improved some layouts
+* Template functions
 * Added info icon to toggle restrictions
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)